Give FFI ability to create "libRebol" GC-safe REBVAL cells
When interfacing Rebol with C, it is useful to be able to reference a
Rebol value through a pointer in a persistent fashion.  Historically
this is a dodgy proposition, because a lot of care must be taken in
the internals of the interpreter to not hold onto a REBVAL* longer than
that address is known to be stable.  Client code could not create a
REBVAL cell in its own address space and expect the GC to know about
it, hence any series pointers held by such values might get garbage
collected out from under it.

Ren-C introduced a new lightweight way of making REBVAL cells that
were GC-visible, but could have flexible lifetime management.  These
API-oriented cells would appear compatible with other cells, hence not
needing a different data type (e.g. RXIARG).  But their underlying
storage would not be in a frame or an array, yet inside a REBSER node
itself.  Their lifetime could be tied to a FRAME!'s lifetime, to a
reference count, manual allocation and freeing, or other concepts due
to the presence of a "meta" value cleverly packed alongside it.

So far RenCpp has been the only client, as the RL_API has not been
retrofitted yet to use these special nodes.  However, an issue came up
in the FFI of needing a similar feature that could be called from
Rebol, as opposed to from C.  This could be used to tunnel Rebol values
through FFI code which desired to persist references (such as
associating Rebol functions with a Tree Control).

The added routines give a manual memory managed way to allocate,
free, set, and get a Rebol value...returned as a platform-pointer-size
integer.  ALLOC-VALUE-POINTER, FREE-VALUE-POINTER, SET-AT-POINTER,
and GET-AT-POINTER.