Remove CLOSURE and `<durable>`, `<has>` back to `<static>`
Ren-C originally aimed to bring the features of CLOSURE! to all
FUNCTION!s at "acceptable cost".  This was partially achieved with
specific binding, and with the ability to use non-reified stack frames
as "specifiers" the costs are hopefully becoming very low for that
binding property (and will be extended to other binding forms).

The second property of CLOSURE! was the indefinite lifetime of all
arguments and locals after the function call.  This was costly, and
entailed making a deep copy of the body of the function on every
invocation.  Ren-C managed to remove the need for that copy, but still
the idea of creating a GC-tracked object for *all* the locals and
arguments in each invocation was heavy-handed.

With implementation techniques getting more sophisticated, the burden
was increasing on having two modalities of function arguments.  Also,
it was seeming less like closure semantics by default was truly a
good idea...the likelihood of obscuring bugs or introducing performance
problems was high.

Rather than dig deeper into CLOSURE and `<durable>` issues for now, it
is likely better to focus on streamlining the code so that methods of
virtual binding can be explored.  In the meantime, USE has been made
into a native, so closure-like effects can be accomplished by putting
USE in the body (as opposed to previously, where it was done in reverse
and USE was written in terms of CLOSURE).

Due to the possibility of `<has>` being coopted for "durable" variables
and given the desire for `<static>` to be retained regardless, this
plays it safe and returns the usages of `<has>` to `<static>`.