Speed bindability check in debug build, remove some cast()s
"Bindability" is whether a REBVAL* is the kind that has a binding to
a context, frame, or function.  Non-bindable values are those which
use the "extra" field in the cell for other purposes (e.g. DATE!).
The REB_XXX types are arranged so that all the bindable types come
before all the non-bindable types, so a separate flag is not needed.

However, using VAL_TYPE() in the debug build to get the REB_XXX and
then do all the associated checks is a bit prohibitive, considering
how often the flag needs to be tested.  This changes to use the
VAL_TYPE_RAW()...because odds are a VAL_TYPE() will come along later
and check that value anyway.

Additionally, this does some tweaking of places that use the cast()
helper which don't really need the protections it gives as much as
they need a little bit more performance considering how often the
routines are called in the C++ debug build.  This does reduce the
"queryability" of the codebase somewhat (the cast constructs allow
for specific casts to be found and studied), and isn't a good general
practice to use raw C casts, but these few cases are probably worth it.