Use simple pointer incrementation Fetch_Next_In_Frame()
Ren-C abstracted the "source" for an evaluation (or for parse rules)
so that the same routines could be used to process either a Rebol
array of values or a C va_list of REBVAL* pointers.  In the process of
this abstraction, it was necessary to ensure that the evaluator was
systematic about feeding forward with one unit of visibility at a time,
since it is not possible to index backwards into a C va_list.  That
was accomplished with a routine called Fetch_Next_In_Frame().

The unusual EVAL operation introduced a requirement to be able to
allow a completely synthesized value to inject itself into the train
of evaluation.  This synthesized value would have to be GC-safe during
the course of its usage, and would have to preempt the value which
would ordinarily come from Fetch_Next_In_Frame().  Hence the idea of
a ->pending value was introduced to the frame, which could either be
NULL or a value that had been deferred and bumped to the pending slot
by the EVAL operation.  This was treated as a generic feature, with
the idea that other operations might want to splice arbitrary values
into the executing source.

With time, needs of enfix processing led the evaluator to have some
specific processing that provides ahook where EVAL's needs can be
met without reaching into Fetch_Next_In_Frame().  This means that the
fetching could be simplified a bit.  (While it makes things slightly
faster, the simplification benefit is probably more significant.)