Move error to separate function to help Get_Var_Core inlining
The inlining of Get_Var_Core wasn't working as well as it should,
apparently because of the presence of a DECLARE_LOCAL inside its body.
This was leading to more instructions just to set up and take down the
function, and since Get_Var functions are called so frequently this
was creating a problem.

This moves the error out to a separate function.  A similar thing is
included to move DECLARE_LOCAL out of Do_Core, but since it's not
inlined there is apparently no effect.  But it's kept just for the
sake of making Do_Core a little bit shorter.