Minor function dispatch optimizations
This takes advantage of the fact that function bodies are always at
index 0 for interpreted functions (at the moment) to avoid fetching the
VAL_INDEX() when the function is dispatched.  It also takes advantage
of the fact that FUNC_BODY is always coming from a singular array to
avoid testing the SERIES_INFO_HAS_DYNAMIC flag.