Simple metrics gathering hook measuring function-call counts
This is a first attempt at building an evaluator hook which builds a
map from FUNCTION! to the first non-anonymous symbol a function looks
up to, and the number of times that function was called.

The interface is a bit clunky.  It is turned on with METRICS ON, and
then later turned off with METRICS OFF...returning a MAP!.  This is a
globally referenced map, which must be cleared between runs.

Sample usage:

    metrics on
    ...your code here...
    m: metrics off
    for-each [fun info] m [
       symbol: info/1
       num-calls: info/2
       print [symbol "was called" num-calls "times"]
    ]
    clear m ;-- before next run