Fix copying of FRAME!s
The code for copying an ANY-CONTEXT! presumed the storage for the vars
of all contexts lived entirely in the varlist.  This is not true for
FRAME!s that are backed by chunk stack memory (e.g. those FRAME!s that
were a by-product of a function call, as opposed to a MAKE FRAME! call
which uses dynamic memory).  This fixes that COPY problem.