Small performance tweak when skipping refinement arguments
The ->refine field of a REBFRM* not only encodes the REBVAL pointer of
where the current refinement is, it also holds special values to
encode the state of traversal of the arguments.  The special values
are chosen so that IS_TRUTHY() can be used to determine whether the
argument gathering loop needs to do type checking.

One value used VOID because it did not have to participate in this
IS_TRUTHY() check, since it indicated skipping refinement arguments
and the code would never get to that point.  However, since it never
gets to that point it is cheaper to pick NULL for this state, since
testing a pointer for NULL is generally faster than testing against
an arbitrary pointer in most compilers.