Streamline and speed up function dispatch in Do_Core()
In early development of Ren-C, it made it possible to re-derive a
4-platform-pointer-sized function REBVAL entirely from a platform
pointer...using a trick much like how OBJECT!s were able to reconstruct
themselves with extra data living in the [0] cell of the array which
held their variables.  Hence, internally to the evaluator a frame's
function was extracted from the value cell and set to a single REBFUN*.

This allowed the code to sidestep the problem of dealing with a
plain REBVAL* which could not be used as the frame's concept of the
currently running function, because that pointer might be into f->out,
and would be overwritten in the course of the evaluation.

However, with the advent of bindings in functions (the way a single
RETURN archetype can have different REBVALs which supplement that
REBFUN* with saying where to return *from*), it was not possible to
just set f->fun to the REBFUN* and overwrite the dispatching REBVAL.
The binding had to be preserved.  The stylization that emerged was
that before jumping to the `do_function` label, a pointer to an
appropriate REBVAL* had to be initialized so that the binding AND
the REBFUN* could be extracted from it.  This was coupled with several
other required steps.

This folds those steps into the reusable inlinable Push_Function(), and
just adds more calls to that.  This way, it's not necessary for the
places that want to push functions to jostle around the function or
binding from their location, they just pass it directly...along with
the label of the frame, and it sets the eval_type to REB_FUNCTION also.

Additionally this commit moves some of the pre-function-processing
assertions into %d-eval.c to shorten the source in Do_Core() a bit.
Some new debug invariants are added as well.