Change f->special for normal argument fulfillment to NULL
f->special indicates the argument mode, whether it be ordinary
fulfillment, specialization, or just enumerating the arguments and
doing type checking with no fulfillment whatsoever.

Indicating ordinary argument fulfillment case previously used the canon
END cell, instead of NULL.  This led to a guarantee that once the
enumeration was over, f->special would always be END, regardless of
which of the three cases it was.  The GC could then also not bother
with testing for NULL.

That invariant is of questionable use, considering how often the
comparison of `f->special == END` has to be done.  Since that test is
run for every step of argument gathering or checking, it's more
efficient to test against NULL, since nearly all compilers can generate
a faster test against 0 than against an arbitrary memory pointer.

Also includes some better documentation about what is going on in
the process_function loop.