Fix undiagnosed longjmp clobber in REMOVE-EACH
Optimized builds revealed a problem with the way REMOVE-EACH was
written, where a local variable was being clobbered, when it was
important to have up-to-date in the event of a longjmp (FAIL).  This
affected interrupted REMOVE-EACH on BINARY! and STRING!.

(For some reason, GCC did not warn about the problem.)

This uses the standard approach of wrapping the setjmp code into its
own function, which builds further upon the already-existing need to
have REMOVE-EACH hold its state in a structure that could be passed
to different finalization routines.

Also makes it so that the `DEBUG=callgrind` build has access to the
stdio routines for printf output.