Don't use a frame when building paramlist from func specs
For reasons that aren't particularly clear, Make_Paramlist was using
a Reb_Frame to enumerate the function spec.  It makes sense for code
like CASE or ALL which are repeatedly calling into the evaluator, and
which also want to localize errors to the places they occur.  But here
it just adds more overhead with no benefit over simple pointer
incrementation.

This takes that out, and also tidies a couple of other frame-related
things, including fixing the debug invariant for f->pending when the
do count breakpoint is used.