Dialectize %systems.r, fold in %platforms.r, add legacy defs
The systems.r table was separate from some partially redundant
information in platforms.r.  This creates a simple dialect to reflect
the builds, that folds the platform information into the same table.
It also uses different data types to make the build setting summaries
easier to distinguish.

At the same time, this brings in historical build information and
numbering...for the sake of discussion.  There are holes and some
inconsistencies, and it's not necessarily desirable to keep all of
them around going forward.