Some reformatting of Travis file for readability, brew update
This adds some comments and organizes the Travis file a bit better,
using multi-line formats for the matrix environment variables.  This
makes it easier to see as well as easier to comment on line-items.

Also attempts to resolve the OS X build failures via `brew update`
before running `brew install`