Auto-sense OBJ paths from source file paths given, vs. guess
The build has a list of source files and what subdirectories they
live in.  %make-make.r was trying to create obj subdirectories to line
up with those files, but did so by examining the "os-base".

This removes that logic, and just walking the list of source files and
adds any relevant subdirectories.