Some fixes pertinent to making Ren Garden work
This includes a bugfix to the code for copying contexts, which was
not updating the length of the generated array correctly.  It also
corrects a spurious assert in the GC when unreadable blanks are used
over the course of time that an evaluation happens (the goal of an
unreadable blank as opposed to trash is that it can survive GC).

Also corrects a compilation error that can arise when linking to
a Rebol built as C to a C++ application, and a C++14 free() warning.
Includes a test to see if the C++ MinGW build reverted to trusty will
result in the debug CPP build working.