Initialize frame "spare" cell in DECLARE_FRAME to unreadable blank
Optimization efforts in the early days of Ren-C attempted to minimize
the number of assembly instructions in each evaluator step, which
included not needing to push protection for the temporary "eval"
cell.  This led to situations where trash would be legal in that cell.

While not all frame fields need to be protected, the cell is being
formatted with a bit pattern anyway (all cells must be now).  That bit
pattern might as well be non-trash, just to simplify the invariant.
This also means that the cell does not have to be initialized on
each subsequent function call, as even if it contains some holdover
temporary garbage it shouldn't be a problem.  The debug build resets it
to unreadable blank on each evaluation step, just to make sure the
temporary garbage isn't used.

(Note: Some imagined concept of security might say letting a native
function potentially examine temporary state from the last evaluation
is a bad idea, but one can't really bulletproof C code in Rebol against
that in any general sense.  It may, however, be desirable for other
additional reasons to SET_END() on the temporary cell before each
function call.  Review.)

Also adds a dump of the current tick count on a panic(), and removes
a possibly spurious assert from the GC.