LIT-WORD! passed as loop variable reuses existing binding
This change allows a LIT-WORD! to be passed to loops, e.g. FOR-EACH,
and signify that an existing variable should be assigned instead of
a new variable and binding.  So:

    >> x: 10 | for-each 'x [1 2 3] [print x] | print x
    1
    2
    3
    3

This contrasts with (and is faster than) WORD!, which generates a new
variable and will currently deep copy and rebind the body to that new
variable:

    >> x: 10 | for-each x [1 2 3] [print x] | print x
    1
    2
    3
    10

(Note that as of time of writing, Red has chosen the LIT-WORD! semantic
for regular WORD!s--a breaking change from historical Rebol.)

If multiple variables are specified, e.g. `for-each [x 'y z] ...`, then
each individual variable will be handled according to the WORD! or
LIT-WORD! status.  Presently this means that if any ordinary WORD! are
used, then the body will be deep copied and rebound.  Future plans
hope to make all aspects of such binding practical and lightweight,
even for deeply nested loops.

Also included in this change is protection against duplicate variables
in a block, e.g. `for-each [x x x] ...` will generate an error.  This
protection is not applied to lit-words like `for-each ['x 'x 'x] ...`
because since they do not represent new bindings, they could each be
bound to different x variables in different contexts.

Additionally, the fabricated object underlying the loop iterations is
protected from expansion...since the loops themselves cache pointers
into the keylist and varlist across arbitrary user evaluations.