LINK() and MISC() macros to access series pointer-size fields
There are two general-purpose pointer-sized fields inside the REBSER
node, ->link and ->misc, used by different series types for purposes
appropriate to the role that series plays.  They have the benefit that
all REBVAL cells which reference the same REBSER node will be able to
inherit updates to those fields (while data stored in the cells
themselves are unique to each REBVAL).

This change helps safely facilitate use of the REBSER nodes themselves
as a way of mapping to their own associated properties (vs. making a
separate hash table of some kind).  This would involve temporarily
moving a series's ->misc field into a cell on the data stack, and then
co-opt the ->misc field to hold that cell's data stack index.  The
series node would be marked with a flag to indicate that it had been
forwarded, and then the rest of the data stack cell's field that wasn't
the saved ->misc pointer would be available for associated properties
(likely up to two additional pointers).  When the process is finished
and the associated pointers no longer needed, the data stack would be
walked and the ->misc pointers restored.

Yet because of the temporary damage, calling a routine which reads
inadvertently from a corrupt ->misc would be bad.  This introduces a
level of macro indirection for both LINK() and MISC(), which should
not add any overhead to the release build...but allows a debug build
to assert if such an attempt is made.

A few cases do not go through this, as it uses SER() which slows down
the C++ debug build in some cases more than seems necessary.

(This is preparatory as a separate commit, as the code that uses this
technique is more clearly seen when kept in its own commit.)