Updates for STRICT_BOOL_COMPILER_TEST
The strict boolean compiler test ensures that nothing but 0 or 1 are
ever passed around in REBOOL, even in old compilers.  It's not done
often because the resulting executable won't run.  This fixes a couple
of violations, does some updates to it so that it works with
RIGOROUS=yes, and simplifies it.

Includes fix for a mistake in the FOR-EACH test.