Disallow RELVAL assignments, introduce Blit_Cell()
Value cells now contain some amount of "formatting" in their headers.
These are some "out-of-band" bits that should be preserved via masking
even when they are being filled with a completely new value.  They
describe aspects of that cell's nature and lifetime which may be
distinct from the characteristics of the cell being copied from.

Because of this, the C++ build stopped allowing even fully specified
REBVAL values to be copied blindly from one spot to another by
arbitrary code.  A special `Move_Value()` operation would have to be
used, that would do the proper masking when moving the header...and
which is able to take C stack pointers and "reify" them into heap
objects if they cross boundaries.

However, some circumstances still were allowing RELVALs to be assigned
to each other.  Some of these were valid (such as moving a cell
between positions in the same array).  Some were invalid.  This removes
the ability to assign RELVALs by default in the C++ build, but also
introduces a special operation called Blit_Cell() which can be used if
the circumstances are right.  It includes a check that the header bits
in the source cell that would be mask protected are identical to the
bits in the target cell, such that it is safe to do.

This also does a light cleanup of the relationship between specific
and relative values, defining a base class Reb_Cell which is distinct
from RELVAL, so that it can be resident in the series content and still
allow the series content to be copied.