Only enforce "Casts for the Masses" in the release C++ build
While the syntax of the `cast(..., ...)` macro is much better than the
traditional C cast, and the checking is useful, it can be very annoying
when trying to step through the debugger and getting into a operation
that serves no purpose after compilation has succeeded.  Also debug
builds aren't optimized, so a noticeable percentage of the runtime
can be spent running something that's a no-op.

This turns off the static analysis behavior of the casts in debug
builds, saving it for just release builds, and promotes one of the
Travis OS X release builds from C to C++ so there's at least one build
checking the rules.

Also fixes a problem in plain C++98 builds.