Make the Error_XXX_Raw functions take REBVALs not RELVALs
The Error() generator is variadic, and attempts to read REBVAL* out of
the void* results of the va_arg access.  If these values have the
relative bit set, they will be asserted upon.  (While it would be
possible to clear out the relativism of an ANY-WORD! and forget its
binding, a relative ANY-ARRAY! would have to be deeply recursed to
do such a scrubbing...otherwise badly structured data would be exposed
to the system.)

The Error_XXX_Raw functions are there to check the arity of error
invocation, but they should also check the REBVAL* vs. RELVAL* status.
Previously they were taking RELVAL...this strengthens them to take
REBVAL.

Includes a small amount of code cleanup on signals and a few comments
about the PNG build settings...as well as cleaning up a bit copy of
value cells in the FFI to use Move_Value().