Begin Streamlining of Collect Words / Typeset code
The recursive walk used mostly for collecting SET-WORD! instances in
a block (for MAKE OBJECT! and FUNCTION) is historically very dodgy
code.  It has been made difficult because of a desire for efficiency
which involved things like not keeping the length of the array where
the words were gathered (BUF_COLLECT) accurate on each gathering,
presumably because of some early micro-optimization during R3-Alpha's
performance analysis.  It was also difficult because a "binding table"
which mapped from canon symbol IDs to their bind indices was static and
global...and all exit paths from the code, from failures or otherwise,
needed to make sure that associations from words to their bind indices
were cleared so as not to corrupt the next bind operation.

Ren-C eliminated the "bind table", instead turning the data backing
ANY-WORD! symbols into series.  Then bits were used inside the series
node itself to maintain the bind table information for that symbol in
effect for the current collection or binding operation.  However, this
had the same issue of this being a kind of global state that had to
be cleared out by all exit paths from the code that used it.

The concerns driving the non-terminated state of the BUF_COLLECT array
make it a good candidate for the code using it to be switched to using
the data stack.  Also, it is desirable to be able to deal with cyclic
structures and not crash.  But improving the code requires it to be
simpler and more obvious.  This introduces the idea of a "collector"
struct which manages the state during a collect, including what index
the next collected word will be at.