SET-META individually on ANY-CONTEXT! (including FRAME!)
There are two spare platform-pointer-size fields in every REBSER node
which are used for various purposes, depending on the role of that
node.  The early development of the "meta" feature would use one of
the fields on the *keylist* of objects to store the pointer to the
ANY-CONTEXT! to serve as the meta object.  But since keylists may be
shared between object instances, this didn't allow different object
instances with shared keylists to have distinct meta object.

This moves the meta link to the varlist to solve that problem.  But
that raised another issue where although FRAME! was an ANY-CONTEXT!, it
was using the varlist slot for an important shared property--namely
the reverse link to the C struct representing an active call frame,
which needed to be NULLed out in the series node so that all FRAME!
REBVALs could tell they had expired.

While it would have been possible to just disallow setting meta-objects
on FRAME!, a relatively elegant solution is to unify the "keylist"
REBARR pointer and the "f" REBFRM pointer into a single field.  A
live frame simply points to the frame, and then decays to the function
of the frame's paramlist when execution has terminated.  Since a REBARR
and a REBFRM can be distinguished by the NODE_FLAG_CELL on the cell
at the head of the Reb_Frame, it's easy to test which situation
applies...which actually increases the performance of CTX_KEYLIST from
its prior implementation.