Move legacy function argument conversion to usermode
The `<r3-legacy>` mode is not really being maintained, and is more of
documentation of what has changed than code people are actively using.
Its performance is not important.  This takes C code which implements
the "blankification" of refinement arguments and moves it into
userspace as part of the legacy wrapper.

It also kills off LIT-WORD-DECAY...an idea which Red has similarly
gotten rid of, and some dead LEGACY(...) options.