Kill off GET_FLAG() and FLAGIT() macros in %reb-c.h
R3-Alpha had a set of generic macros for shifting a numbered bit into
positions, and then setting/getting/clearing that bit.  On the plus
side, this allowed the flags themselves to be declared in an enum
without shifting them.  On the negative side, *almost everything else*.
It made it less clear to work with them at the callsites.

    flags = (1<<FLAG_ONE | 1<<FLAG_TWO);
    if (GET_FLAG(flags, FLAG_ONE)) ...

vs. if the shift were in the enum itself:

    flags = FLAG_ONE | FLAG_TWO;
    if (flags & FLAG_ONE) ...

It's also harder to combine this form of flag, so rather than wind up
with a generic `ALL_FLAGS(x, FLAG_ONE | FLAG_TWO | FLAG_THREE)` macro,
there were versions that took N macro parameters to clear or set bits
together, like `GET_FLAGS(x, FLAG_ONE, FLAG_TWO)`.  This led several
usages to reach "below the abstraction" to do their setting or testing.

More generally, the problem with such a generalized macro is that
because it was working with unknown bit flags there were no real
useful benefits to the abstraction.

(Contrast this with GET_VAL_FLAG() in Ren-C, which takes a whole REBVAL
as opposed to just its flag bits.   This enables various checks, such
as making sure that a flag only applicable to words is not being tested
on an OBJECT!.)

So these bit testing and flipping macros offered no concrete benefit,
and generally just obscured the simplicity of what was going on--which
would be readable and apparent to any C programmer.  They also went
against the conventions used in the rest of the codebase, where the
flags themselves could be OR'd together to initialize them, and tests
could be done without shifting.

However, it's reasonable to imagine that any given abstraction would
have its own specialized macros for testing flags--but those would
hopefully add value.  Hence abstractions like GET_GOB_FLAG() were left
in, since they take an entire gob--not just the flags extracted from
the gob.