Move most REBXXX definitions out of %reb-c.h to %reb-defs.h
The file %reb-c.h is supposed to be a set of generic C routines that
might be useful to anyone, whereas %reb-defs.h has evolved to be the
common include file for both %sys-core.h and %reb-host.h for the
definitions of REBXXX types.

Since the abstract definitions of integers like u8, u32, and u64 is
sometimes all files need to use (without knowing if REBCNT is 32-bit
or 64-bit... it may someday become platform dependent) it's best to
split them off.

This also gets rid of some non-inline functions that were used prior to
the C++ build to do type checking, instead favoring inline functions.