Make source scanning accept a va_list for splicing in REBVALs
Because of the careful choices made in arranging the bits of headers of
Rebol values, it's possible to know from examining the first byte of
a void pointer whether it is a UTF-8 string or a REBVAL*.

This adapts the scanner to be able to accept a feed of pointers to
UTF-8 strings and Rebol values (as opposed to just one UTF-8 pointer).
If a value is encountered in-between tokens, that value will be
"spliced" into the scan.

Currently this can be used to construct arrays, such as in this code:

    REBVAL *item1 = ...;
    REBVAL *item2 = ...;
    REBVAL *item3 = ...;

    REBSTR *filename = ...; // where to say code came from

    REBARR *result = Scan_Va_Managed(filename,
        "if not", item1, "[\n",
            item2, "| print {Close brace separate from content}\n",
        "] else [\n",
            item3, "| print {Close brace with content}]\n",
        END
    );

While this concept has several applications, the most obvious is in the
creation of a friendly C API for constructing code.