GET-ENV/SET-ENV => process extension, RL_XXX callable from Core
This takes a step toward moving the RL_API forward and eliminating
REBCHR from the API, by removing the split between the environment
reading and writing functions and the "host".  Instead of needing to
go through an intermediate representation, the conditional code on
Windows vs. POSIX handles REBVAL* directly...extracting wide characters
from the REBVAL via the RL_API or UTF8, as appropriate based on the
platform.

Because the model being pursued in the new RL_API is actually likely
to be useful to the core at times, this sorts out some of the #include
problems which hampered calling the RL_XXX functions when one was
including %sys-core.h