Move REQUEST-FILE into extension, patch it for Windows/GTK
The REQUEST-FILE dialog was implemented in GTK and Windows in the
Saphirion/Atronix repositories.  Because they originated from times
before Rebol was open sourced, they had to implement their features in
a rather circuitous way: providing "host services" that didn't speak
any Rebol datatypes and awkward COMMAND!s to hook to the interpreter.

Following the model of simplification that has been evolving, this
puts the platform-specific code into an extension--tentatively the
"View" extension.  This brings the code up to date somewhat and tests
it (some) on both Windows and GTK.  It eliminates the REBRFR ("Rebol
File Request") structure, and allows the implementations to go more
directly to the formats they need.  (e.g. if the result is a BLOCK!,
to be able to construct that block from the data instead of going
through an abstraction of a stringlist, just to make Windows and GTK
force through a common arbitrary point).

This also breaks out the REQUEST-DIR code, which introduced an added
dependency on OLE32 since it uses a somewhat obscure shell API instead
of a typical "common dialog".  While the code now works, it's not
necessarily that great a dialog, and it seems the common dialogs can
handle directory selection...just not as easily (but the work has
already been done).