Remove Make_OS_Error() from core
Turning an OS error into a Rebol ERROR! is a generally useful routine.
However, the goal is to make the core platform independent code (with
no #ifdef TO_WINDOWS etc.)  Because it was not platform indendent,
the error API had to speak in terms of "REBCHR", which is being
eliminated.

Since the only current client of that routine was in the Process module
the un-hostkitted code was unified and moved there.  It raises the
question of how such useful C routines might be shared across
extensions that need them, but that could be done other ways (such as
a common .inc file, even.)