Get rid of proto-parser "style" field
For a time during its development, the proto-parser was being written
to work on either new or old format headers.  The old format header
code was removed, but the `style` field was not.  This made the code
slightly more difficult to comprehend...when being able to tolerate
more than one function header style at a time is not a feature that
the system truly needs.  (The move to using Rebol generically in those
function headers is considered "final".)

Also adds some comments on END based on a failed experiment to turn it
into a C string literal (explains why that won't work in today's code)
and renames %reb-lib-lib.h to the more accurate %tmp-reb-lib-table.inc