Rename RL_XxxYyy external API to rebXxxYyy, emitter objects
This switches the current grab bag of "libRebol" functions to use the
general naming pattern of libRed.

What made this non-trivial is that if Rebol is built as an EXE instead
of a DLL (as it usually has been) then there is no standard C way of
loading an extension DLL linked directly against exports from the EXE.
Hence a table of functions has to be passed to the extension.  Calling
functions from the table requires macros.

Code for doing this was part of hostkit, but it has been difficult to
read and modify easily.  Also, it did not provide macros which allowed
code from inside the codebase to invoke the APIs under the same name
as the code outside.  This tries to bring the code under more control
for the change, and paves the way to using multiple "emitters" at the
same time with their own buffers...which might allow the generative
tools to do more in one pass.