First attempt to use more "libRebol" in REPL C code (leaks)
This is a semi-working implementation of a rebDo() API which is
sufficient to implement the C portion of the REPL code.  It comments
out some useful asserts in order to get the evaluator to allow direct
evaluation into API cells.  While it shows promise, a lot of new code
will have to be written before it's ready for prime-time.