Change to NULL signal on f->value when frame ends vs. END
Originally a FRAME! could be based on either feeding a C va_list of
values (as called from C, without needing to create an array) or an
array (as invoked by something like DO or any other array-based
executor).  It was one or the other.

A goal of the rebDo() API is to mix these approaches, to where "runs"
of values can come from UTF-8 strings in with ordinary REBVAL* that
have been spliced in inertly (or brought to life with rebEval()).
To achieve this, reaching an END has to be a signal not to end the
input, but to potentially consult a va_list for the next value or
run of values to process.

This ceases the use of END as the trigger for when a frame is at its
end, rather seeing if the fetching process returns a NULL.  This is
not strictly necessary as an END internal to the fetching process
need not return that END, but it creates a clear distinct state which
will be helpful in being confident in the fetch.