Migrate frame's index/pending fields into the source struct
This takes a further step toward making a potentially compatible
interface between the variadic scanner and a va_list based frame.  It
makes sure direct accesses to the va_list expecting REBVAL* only are
not used, but always defer to Fetch_Next_In_Frame()...which can use
pointer detection to see the UTF-8 case and defer to the scanner.

Also tidies up several of the entry points that wrap up Do_Core so
that they are more clear and contain less outdated information.