Convert rebEval() to an "eval instruction" instead of pure REBVAL
This resolves issues of memory leaking in the fledgling libRebol-based
REPL code for C.  It changes rebEval() so that it uses a singular array
to specify eval instructions.  Those instructions are explicitly
visited by the variadic feeding process for Fetch_Next_In_Frame(),
which moves them into the frame's spare cell and could (theoretically)
free the instruction at time of visitation.

While this could be made to work freeing the instruction upon the
visit, a safety assertion currently stops it because it appears a
manually-created series is being freed at the wrong time.  That is an
important assert--more important than deferring for the moment and
just making the instructions GC-managed for now.

Some reorganization will be needed before the preloading of a frame can
handle non-REBVAL pointers, so at the moment a BLANK_VALUE is seeded in
the variadic first slot for code that wants to do a rebEval() as its
first action.