Fix bug affecting routines like TRIM with high codepoints
SET_ANY_CHAR was originally a macro in R3-Alpha.  But it repeated its
arguments, so it could trigger multiple evaluations.

When it was converted to an inline function in Ren-C, it was
incorrectly given an argument for the character as REBYTE instead of
REBUNI, causing high codepoints to be truncated to <= 255.  This broke
routines like TRIM for those high codepoints.