Only duplicate ANY-ARRAY! deeply during bind/loop bodies
R3-Alpha's loops (and still, often the loops in Ren-C today) which
have loop variables must bind their bodies to a new context.  In order
to not disrupt the bindings of any passed-in block data, the loop
bodies are copied.  However, an apparently longstanding inefficiency
would lead it to copy not only the ARRAY! values but also string
series.

Even if ANY-STRING! could have bindings that would be affected, those
bindings would live in the value cells themselves...not the series.
This addresses the inefficiency by changing those copies (also used by
USE and BIND/COPY) to only copy the array backbone for the value cells.

(Note: virtual binding aspires to be even less inefficient, where no
copies of the array backbone are needed either.)