POSIX console UTF8, multiline paste, escape/cancel INPUT
The very old/crufty R3-Alpha POSIX termios-based code was considered to
be only a temporary codebase to be replaced with something better.  No
truly suitable replacement has been found, so this commit bites the
bullet and pushes the existing code to make it more tolerable.

It now supports UTF-8 entry...although the cursor-oriented commands
need to be updated to account correctly for multibyte characters.
Though those cursor commands are unfinished, it is still a significant
improvement.

It also supports multiline pasting.  This was inspired by a patch
submitted by @Oldes as a proof of concept (and the provocation for
this reworking).  Consequently hitting enter on an empty line no longer
terminates in-progress multiline input with an error message.

The way to cancel REPL input is now to use Ctrl-D on POSIX and Windows,
or Escape on POSIX.  (The technical reasons why Windows cannot use
Escape are explained in detail in the commit.)  The reaction of the
INPUT native is distinct between Ctrl-C and Ctrl-D/Escape, because
it's the difference between halting the script code completely, or
just returning BLANK! from INPUT and allowing the script to make a
decision on what to do from there.

Unknown escape codes are handled more intelligently, by not letting
stray bytes loose into the input buffer.  In the future, the escape
codes should be bubbling up to usermode somehow, to make it easier
for user customization and so there's less C code overall.

And importantly: the code is restructured and annotated to make it
easier for interested parties to contribute.